+++ inc/settings-page-admin.php
@@ -206,6 +206,7 @@ function elodin_bridge_render_admin_page() {
 	$editor_ui_restrictions_enabled = elodin_bridge_is_editor_ui_restrictions_enabled();
 	$editor_publish_sidebar_restriction_enabled = elodin_bridge_is_editor_publish_sidebar_restriction_enabled();
 	$editor_show_template_default_enabled = elodin_bridge_is_editor_show_template_default_enabled();
+	$editor_list_view_default_enabled = elodin_bridge_is_editor_list_view_default_enabled();
 	$media_library_infinite_scrolling_enabled = elodin_bridge_is_media_library_infinite_scrolling_enabled();
 	$shortcodes_enabled = elodin_bridge_is_shortcodes_enabled();
 	$svg_uploads_enabled = elodin_bridge_is_svg_uploads_enabled();